FIXED: Fix PATH in convenience scripts
Add the convenience script's directory to PATH, on the assumption that
the multimarkdown executable is in the same directory.

Remove $PWD from PATH, because there's no reason why we should expect
the user's shell to have been changed to the directory that contains
multimarkdown.

export PATH separately from setting PATH, because older shells can't do
both at once.

unset CDPATH because if a user had exported it, it could change the
behavior of cd.